Partially split up chainlink declaration files
* Babel cant handle multiple mappings to the same
* declaration file, so we have to split up our declarations a bit
